requirements.txt
altgraph==0.17.2
colorama==0.4.5
future==0.18.2
loguru==0.6.0
macholib==1.16
pefile==2022.5.30
psutil==5.9.1
pycryptodome==3.15.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pywin32-ctypes==0.2.0
tinyaes==1.0.3
win32-setctime==1.1.0
